+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.27.7
+
+Generic Form
+
+- Now provides pickers for numerical color fields
+- A function can be passed as 'inputChangeCallback' to return selected fields every time form input changes
+
 # 1.27.5
 
 Generic Forms